chore: Setup Basic Tabs
