+++ .gitignore
@@ -55,7 +55,7 @@ buck-out/
 */fastlane/Preview.html
 */fastlane/screenshots
 
-storybook-native/story-loader.js
+app/storybook/story-loader.js
 coverage
 dist
 packages/*/rnw.js
