+++ app/.eslintignore
@@ -1,0 +1,2 @@
+storybook/story-loader.js
+metro.config.js
