+++ app/.prettierignore
@@ -1,0 +1,1 @@
+storybook/story-loader.js
