+++ app/navigation/ids.js
@@ -1,0 +1,5 @@
+export const TAB_ID_HOME = "HOME_TAB";
+export const TAB_ID_MORE = "MORE_TAB";
+
+export const VIEW_ID_EDITION = "thetimes.views.FirstView";
+export const VIEW_ID_STORYBOOK = "thetimes.views.Storybook";
