+++ app/navigation/index.js
@@ -1,0 +1,11 @@
+import { Navigation } from "react-native-navigation";
+
+import * as ViewsIDs from "./ids";
+import { EditionView } from "../views/Edition";
+import { StorybookUIRoot } from "../storybook/";
+
+// prettier-ignore
+export const registerScreens = () => {
+    Navigation.registerComponent(ViewsIDs.VIEW_ID_EDITION, () => EditionView);
+    Navigation.registerComponent(ViewsIDs.VIEW_ID_STORYBOOK, () => StorybookUIRoot);
+};
