+++ package.json
@@ -17,7 +17,7 @@
     "storybook:build": "yarn storybook:build-vendor && build-storybook -c .storybook -s ./dist/public",
     "prestorybook:test-build": "yarn storybook:build-vendor",
     "storybook:test-build": "build-storybook -c .storybook -s ./dist/public",
-    "storybook-native": "react-native start --config ./storybook-native/metro.config.js",
+    "storybook-native": "react-native start --config ./app/metro.config.js",
     "prestorybook-native": "rnstl",
     "ios:app": "react-native start --config ./lib/ios/metro.config.js",
     "ios:build-lib": "yarn ios:build-lib:js && yarn ios:build-lib:update-version",
@@ -58,7 +58,7 @@
         "packages"
       ],
       "pattern": "**/*.stories.js",
-      "outputFile": "./storybook-native/story-loader.js"
+      "outputFile": "./app/storybook/story-loader.js"
     }
   },
   "homepage": "https://github.com/newsuk/times-components-native#readme",
