--- storybook-native/.eslintignore
@@ -1,2 +1,0 @@
-story-loader.js
-metro.config.js
