--- storybook-native/.prettierignore
@@ -1,1 +1,0 @@
-story-loader.js
